requirements.txt
black==22.3.0
click==8.1.3
inquirerpy==0.3.3
mypy-extensions==0.4.3
paho-mqtt==1.6.1
pathspec==0.9.0
pfzy==0.3.4
platformdirs==2.5.2
prompt-toolkit==3.0.29
pycodestyle==2.8.0
Pygments==2.12.0
questionary==1.10.0
regex==2022.6.2
six==1.16.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
wcwidth==0.2.5
